fix(Desktop): Files which use both windows and unix separators are now visualizable.
To get path to one document, we concatenate a root path defined locally with the document path defined in the form. If those two variables are defined in different systems, it would cause an error when the full path is accessed. Also, the button which opens files on system is now accessible even in consultation mode.