fix(desktop): Fix saving problems on SR edition from map tool. Improve table sorting.
The tool was designed to save data on close. But here were two flaws : First, we tested only troncon state in order to determine if we should save SR or not. Second, when closing using the button in bottom-right corner, no close request event was launched, and so no saving operation was triggered.

About sorting : Tool was sorting bornes from uphill to downhill. The flaw here is that it was a one shot sort, not refreshed on list change. Second flaw was that user did not have the possibility to sort by PR or name. We added those features as well.

JIRA SYM-1430